+++ texlite/components/common.py
@@ -7,6 +7,7 @@ BANNER_LINE = r'% ' + '-' * 78
 SPECIAL_CHARS = r'\#$%^&_{}~'
 NON_ENCAPSULATION_CHARS = SPECIAL_CHARS + r' '
 FONT_SIZES = ['8pt', '9pt', '10pt', '11pt', '12pt', '14pt', '17pt', '20pt']
+EMPTY_LINES = ['', r'\t']
 
 # text regex patterns
 BOLD_RE = r'\*\*([^\*\*]*)\*\*'